Remove old metrics code
