+++ discovery-provider/integration_tests/queries/test_get_aggregate_route_metrics.py
@@ -12,7 +12,7 @@ from src.models.metrics.aggregate_monthly_total_users_metrics import (
 from src.models.metrics.aggregate_monthly_unique_users_metrics import (
     AggregateMonthlyUniqueUsersMetric,
 )
-from src.queries.get_route_metrics import _get_aggregate_route_metrics
+from src.queries.get_aggregate_route_metrics import _get_aggregate_route_metrics
 from src.utils.db_session import get_db
 
 limit = 2
