+++ discovery-provider/src/api/v1/metrics.py
@@ -1,34 +1,30 @@
 import logging
 
-from flask_restx import Namespace, Resource, fields, inputs, reqparse
+from flask_restx import Namespace, Resource, fields, reqparse
 from src.api.v1.helpers import (
     DescriptiveArgument,
     abort_bad_path_param,
     abort_bad_request_param,
     format_limit,
     make_response,
-    parse_bool_param,
     parse_unix_epoch_param,
     parse_unix_epoch_param_non_utc,
     success_response,
 )
 from src.queries.get_app_name_metrics import (
     get_aggregate_app_metrics,
-    get_app_name_metrics,
     get_historical_app_metrics,
 )
-from src.queries.get_app_names import get_app_names
 from src.queries.get_genre_metrics import get_genre_metrics
 from src.queries.get_plays_metrics import get_plays_metrics
-from src.queries.get_route_metrics import (
-    get_aggregate_route_metrics,
-    get_historical_route_metrics,
-    get_route_metrics,
+from src.queries.get_historical_route_metrics import (
+    get_historical_route_metrics
 )
+from src.queries.get_aggregate_route_metrics import {
+    get_aggregate_route_metrics
+}
 from src.queries.get_trailing_metrics import (
     get_aggregate_route_metrics_trailing_month,
-    get_monthly_trailing_route_metrics,
-    get_trailing_app_metrics,
 )
 from src.utils.redis_cache import cache
 from src.utils.redis_metrics import (
@@ -39,13 +35,8 @@ from src.utils.redis_metrics import (
 )
 
 from .models.metrics import (
-    app_name,
-    app_name_metric,
-    app_name_trailing_metric,
     genre_metric,
-    plays_metric,
-    route_metric,
-    route_trailing_metric,
+    plays_metric
 )
 
 logger = logging.getLogger(__name__)
@@ -53,23 +44,6 @@ logger = logging.getLogger(__name__)
 
 ns = Namespace("metrics", description="Metrics related operations")
 
-route_metrics_response = make_response(
-    "metrics_reponse", ns, fields.List(fields.Nested(route_metric))
-)
-route_metrics_trailing_month_response = make_response(
-    "route_metrics_trailing_month_response", ns, fields.Nested(route_trailing_metric)
-)
-app_name_response = make_response(
-    "app_name_response", ns, fields.List(fields.Nested(app_name))
-)
-app_name_trailing_response = make_response(
-    "app_name_trailing_response",
-    ns,
-    fields.List(fields.Nested(app_name_trailing_metric)),
-)
-app_name_metrics_response = make_response(
-    "app_name_metrics_response", ns, fields.List(fields.Nested(app_name_metric))
-)
 plays_metrics_response = make_response(
     "plays_metrics", ns, fields.List(fields.Nested(plays_metric))
 )
@@ -225,109 +199,6 @@ class AggregateAppMetricsTrailing(Resource):
         return response
 
 
-@ns.route("/routes", doc=False)
-class RouteMetric(Resource):
-    @ns.expect(metrics_route_parser)
-    @ns.doc(
-        id="""Route Metrics""",
-        params={
-            "path": "Request Path",
-            "query_string": "Query String",
-            "start_time": "Start Time in Unix Epoch",
-            "exact": "Exact Path Query Match",
-            "limit": "Limit",
-            "bucket_size": "Bucket Size",
-            "version": "API Version Query Filter",
-        },
-        responses={200: "Success", 400: "Bad request", 500: "Server error"},
-    )
-    @ns.marshal_with(route_metrics_response)
-    @cache(ttl_sec=3 * 60 * 60)
-    def get(self):
-        """Get the route metrics"""
-        args = metrics_route_parser.parse_args()
-        if args.get("limit") is None:
-            args["limit"] = 168
-        else:
-            args["limit"] = min(args.get("limit"), 168)
-
-        if args.get("bucket_size") is None:
-            args["bucket_size"] = "hour"
-        if args.get("bucket_size") not in valid_date_buckets:
-            abort_bad_request_param("bucket_size", ns)
-
-        try:
-            args["start_time"] = parse_unix_epoch_param(args.get("start_time"), 0)
-        except Exception:
-            abort_bad_request_param("start_time", ns)
-
-        if args.get("exact") is not None:
-            args["exact"] = parse_bool_param(args.get("exact"))
-            if args.get("exact") is None:
-                abort_bad_request_param("exact", ns)
-        else:
-            args["exact"] = False
-
-        args["path"] = args.get("path") if args.get("path") is not None else ""
-        route_metrics = get_route_metrics(args)
-        response = success_response(route_metrics)
-        return response
-
-
-metrics_app_name_list_parser = reqparse.RequestParser()
-metrics_app_name_list_parser.add_argument("start_time", required=False, type=int)
-metrics_app_name_list_parser.add_argument("limit", required=False, type=int)
-metrics_app_name_list_parser.add_argument("offset", required=False, type=int)
-metrics_app_name_list_parser.add_argument(
-    "include_unknown", required=False, type=inputs.boolean
-)
-
-
-@ns.route("/routes/trailing/month", doc=False)
-class RouteMetricsTrailingMOnth(Resource):
-    @ns.marshal_with(route_metrics_trailing_month_response)
-    @cache(ttl_sec=30 * 60)
-    def get(self):
-        """Gets trailing month route metrics from matview"""
-        metrics = get_monthly_trailing_route_metrics()
-        response = success_response(metrics)
-        return response
-
-
-@ns.route("/app_name", doc=False)
-class AppNameListMetrics(Resource):
-    @ns.doc(
-        id="""Get App Names""",
-        params={
-            "offset": "Offset",
-            "limit": "Limit",
-            "start_time": "Start Time in Unix Epoch",
-            "include_unknown": "Whether or not to include unknown apps",
-        },
-        responses={200: "Success", 400: "Bad request", 500: "Server error"},
-    )
-    @ns.expect(metrics_app_name_list_parser)
-    @ns.marshal_with(app_name_response)
-    @cache(ttl_sec=30 * 60)
-    def get(self):
-        """List all the app names"""
-        args = metrics_app_name_list_parser.parse_args()
-        if args.get("limit") is None:
-            args["limit"] = 100
-        else:
-            args["limit"] = min(args.get("limit"), 100)
-        if args.get("offset") is None:
-            args["offset"] = 0
-        try:
-            args["start_time"] = parse_unix_epoch_param(args.get("start_time"), 0)
-        except Exception:
-            abort_bad_request_param("start_time", ns)
-
-        app_names = get_app_names(args)
-        response = success_response(app_names)
-        return response
-
-
 valid_trailing_time_periods = ["week", "month", "all_time"]
 trailing_app_name_parser = reqparse.RequestParser(argument_class=DescriptiveArgument)
 trailing_app_name_parser.add_argument(
@@ -341,67 +212,6 @@ trailing_app_name_parser.add_argument(
     location="path",
     description="The trailing time period",
 )
-
-
-@ns.route("/app_name/trailing/<string:time_range>", doc=False)
-@ns.expect(trailing_app_name_parser)
-class TrailingAppNameMetric(Resource):
-    @ns.doc(
-        id="""Get Trailing App Name Metrics""",
-        description="""Gets the trailing app name metrics""",
-    )
-    @ns.marshal_with(app_name_trailing_response)
-    @cache(ttl_sec=3 * 60 * 60)
-    def get(self, time_range):
-        parsed = trailing_app_name_parser.parse_args()
-        args = {"limit": parsed.get("limit", 10), "time_range": time_range}
-        metrics = get_trailing_app_metrics(args)
-        response = success_response(metrics)
-        return response
-
-
-metrics_app_name_parser = reqparse.RequestParser()
-metrics_app_name_parser.add_argument("start_time", required=False, type=int)
-metrics_app_name_parser.add_argument("limit", required=False, type=int)
-metrics_app_name_parser.add_argument("bucket_size", required=False)
-
-
-@ns.route("/app_name/<string:app_name>", doc=False)
-class AppNameMetric(Resource):
-    @ns.doc(
-        id="""Get Metrics by App Name""",
-        params={
-            "start_time": "Start Time in Unix Epoch",
-            "limit": "Limit",
-            "bucket_size": "Bucket Size",
-        },
-        responses={200: "Success", 400: "Bad request", 500: "Server error"},
-    )
-    @ns.expect(metrics_app_name_parser)
-    @ns.marshal_with(app_name_metrics_response)
-    @cache(ttl_sec=30 * 60)
-    def get(self, app_name):
-        """Get the app name metrics"""
-        args = metrics_app_name_parser.parse_args()
-
-        if args.get("limit") is None:
-            args["limit"] = 168
-        else:
-            args["limit"] = min(args.get("limit"), 168)
-
-        try:
-            args["start_time"] = parse_unix_epoch_param(args.get("start_time"), 0)
-        except Exception:
-            abort_bad_request_param("start_time", ns)
-
-        if args.get("bucket_size") is None:
-            args["bucket_size"] = "hour"
-        if args.get("bucket_size") not in valid_date_buckets:
-            abort_bad_request_param("bucket_size", ns)
-
-        app_name_metrics = get_app_name_metrics(app_name, args)
-        response = success_response(app_name_metrics)
-        return response
 
 
 metrics_plays_parser = reqparse.RequestParser()
