--- discovery-provider/src/models/metrics/route_metrics_month_matview.py
@@ -1,11 +1,0 @@
-from sqlalchemy import BigInteger, Column, DateTime, Table
-from src.models.base import Base
-
-# Materialized view
-t_route_metrics_month_bucket = Table(
-    "route_metrics_month_bucket",
-    Base.metadata,
-    Column("unique_count", BigInteger),
-    Column("count", BigInteger),
-    Column("time", DateTime),
-)
