+++ discovery-provider/src/queries/get_app_name_metrics.py
@@ -1,5 +1,4 @@
 import logging
-import time
 from datetime import date, timedelta
 
 from sqlalchemy import asc, desc, func
@@ -10,7 +9,6 @@ from src.models.metrics.aggregate_daily_app_name_metrics import (
 from src.models.metrics.aggregate_monthly_app_name_metrics import (
     AggregateMonthlyAppNameMetric,
 )
-from src.models.metrics.app_name_metrics import AppNameMetric
 from src.utils import db_session
 
 logger = logging.getLogger(__name__)
@@ -154,53 +152,3 @@ def _get_aggregate_app_metrics(session, time_range, limit):
         raise exceptions.ArgumentError("Invalid time_range")
 
     return [{"name": item[0], "count": item[1]} for item in query]
-
-
-def get_app_name_metrics(app_name, args):
-    """
-    Returns the usage metrics for a specified app_name
-
-    Args:
-        app_name: string The name of the app to query for metrics
-        args: dict The parsed args from the request
-        args.start_time: date The date to start the query from
-        args.limit: number The max number of metrics to return
-        args.bucket_size: string A date_trunc operation to aggregate timestamps by
-
-    Returns:
-        Array of dictionaries with the timestamp, count, and unique_count
-    """
-    db = db_session.get_db_read_replica()
-    with db.scoped_session() as session:
-        return _get_app_name_metrics(session, app_name, args)
-
-
-def _get_app_name_metrics(session, app_name, args):
-    metrics = (
-        session.query(
-            func.date_trunc(args.get("bucket_size"), AppNameMetric.timestamp).label(
-                "timestamp"
-            ),
-            func.sum(AppNameMetric.count).label("count"),
-            func.count(AppNameMetric.ip.distinct()).label("unique_count"),
-        )
-        .filter(
-            AppNameMetric.application_name == app_name,
-            AppNameMetric.timestamp > args.get("start_time"),
-        )
-        .group_by(func.date_trunc(args.get("bucket_size"), AppNameMetric.timestamp))
-        .order_by(desc("timestamp"))
-        .limit(args.get("limit"))
-        .all()
-    )
-
-    metrics = [
-        {
-            "timestamp": int(time.mktime(m[0].timetuple())),
-            "count": m[1],
-            "unique_count": m[2],
-        }
-        for m in metrics
-    ]
-
-    return metrics
