--- discovery-provider/src/queries/get_app_name_metrics_unit_test.py
@@ -1,2 +1,0 @@
-def test():
-    """See /integration_tests/test_get_app_name_metrics.py"""
