+++ discovery-provider/src/tasks/index_metrics.py
@@ -1,11 +1,8 @@
 import json
 import logging
-import time
 from datetime import datetime, timedelta
 
 import requests
-from src.models.metrics.app_name_metrics import AppNameMetric
-from src.models.metrics.route_metrics import RouteMetric
 from src.queries.update_historical_metrics import (
     update_historical_daily_app_metrics,
     update_historical_daily_route_metrics,
@@ -22,15 +19,10 @@ from src.utils.prometheus_metric import (
 from src.utils.redis_metrics import (
     METRICS_INTERVAL,
     datetime_format_secondary,
-    get_rounded_date_time,
     get_summed_unique_metrics,
     merge_app_metrics,
     merge_route_metrics,
-    metrics_applications,
-    metrics_prefix,
-    metrics_routes,
     metrics_visited_nodes,
-    parse_metrics_key,
     persist_summed_unique_counts,
     personal_app_metrics,
     personal_route_metrics,
@@ -39,130 +31,6 @@ from src.utils.redis_metrics import (
 logger = logging.getLogger(__name__)
 
 discovery_node_service_type = bytes("discovery-node", "utf-8")
-
-
-def process_route_keys(session, redis, key, ip, date):
-    """
-    For a redis hset storing a mapping of routes to the number of times they are hit,
-    parse each key out into the version, path, and query string.
-    Create a new entry in the DB for the each route.
-    """
-    try:
-        route_metrics = []
-        routes = redis.hgetall(key)
-        for key_bstr in routes:
-            route = key_bstr.decode("utf-8").strip("/")
-            val = int(routes[key_bstr].decode("utf-8"))
-
-            version = "0"  # default value if version is not present
-            path = route
-            query_string = None
-
-            route_subpaths = route.split("/")
-
-            # Extract the version out of the path
-            if route_subpaths[0].startswith("v") and len(route_subpaths[0]) > 1:
-                version = route_subpaths[0][1:]
-                path = "/".join(route_subpaths[1:])
-
-            # Extract the query string out of the path
-            route_query = path.split("?")
-            if len(route_query) > 1:
-                path = route_query[0]
-                query_string = route_query[1]
-            route_metrics.append(
-                RouteMetric(
-                    version=version,
-                    route_path=path,
-                    query_string=query_string,
-                    count=val,
-                    ip=ip,
-                    timestamp=date,
-                )
-            )
-
-        if route_metrics:
-            session.bulk_save_objects(route_metrics)
-        redis.delete(key)
-    except Exception as e:
-        raise Exception(f"Error processing route key {key} with error {e}") from e
-
-
-def process_app_name_keys(session, redis, key, ip, date):
-    """
-    For a redis hset storing a mapping of app_name usage in request parameters to count,
-    Create a new entry in the DB for each app_name.
-    """
-    try:
-        app_name_metrics = []
-        app_names = redis.hgetall(key)
-        for key_bstr in app_names:
-            app_name = key_bstr.decode("utf-8")
-            val = int(app_names[key_bstr].decode("utf-8"))
-
-            app_name_metrics.append(
-                AppNameMetric(
-                    application_name=app_name, count=val, ip=ip, timestamp=date
-                )
-            )
-        if app_name_metrics:
-            session.bulk_save_objects(app_name_metrics)
-        redis.delete(key)
-
-    except Exception as e:
-        raise Exception(f"Error processing app name key {key} with error {e}") from e
-
-
-def sweep_metrics(db, redis):
-    """
-    Move the metrics values from redis to the DB.
-
-    Get all the redis keys with the metrics prefix,
-    parse the key to get the timestamp in the key.
-    If it is before the current time, then process the redis hset.
-    """
-    with db.scoped_session() as session:
-        for key_byte in redis.scan_iter(f"{metrics_prefix}:*"):
-            key = key_byte.decode("utf-8")
-            try:
-                parsed_key = parse_metrics_key(key)
-
-                if parsed_key is None:
-                    raise KeyError(
-                        f"index_metrics.py | Unable to parse key {key} | Skipping process key"
-                    )
-                source, ip, key_date = parsed_key
-
-                current_date_time = get_rounded_date_time()
-
-                if key_date < current_date_time:
-                    if source == metrics_routes:
-                        process_route_keys(session, redis, key, ip, key_date)
-                    elif source == metrics_applications:
-                        process_app_name_keys(session, redis, key, ip, key_date)
-            except KeyError as e:
-                logger.warning(e)
-                redis.delete(key)
-            except Exception as e:
-                logger.error(e)
-                redis.delete(key)
-
-
-def refresh_metrics_matviews(db):
-    with db.scoped_session() as session:
-        start_time = time.time()
-        logger.info("index_metrics.py | refreshing metrics matviews")
-        session.execute("REFRESH MATERIALIZED VIEW route_metrics_day_bucket")
-        session.execute("REFRESH MATERIALIZED VIEW route_metrics_month_bucket")
-        session.execute("REFRESH MATERIALIZED VIEW route_metrics_trailing_week")
-        session.execute("REFRESH MATERIALIZED VIEW route_metrics_trailing_month")
-        session.execute("REFRESH MATERIALIZED VIEW route_metrics_all_time")
-        session.execute("REFRESH MATERIALIZED VIEW app_name_metrics_trailing_week")
-        session.execute("REFRESH MATERIALIZED VIEW app_name_metrics_trailing_month")
-        session.execute("REFRESH MATERIALIZED VIEW app_name_metrics_all_time")
-        logger.info(
-            f"index_metrics.py | refreshed metrics matviews in: {time.time()-start_time} sec"
-        )
 
 
 def get_metrics(endpoint, start_time):
@@ -356,8 +224,11 @@ def update_historical_metrics(
     update_historical_monthly_app_metrics(db, monthly_app_metrics)
 
 
-# get historical monthly metrics and last month's daily metrics to periodically synchronize metrics across all nodes
 def synchronize_all_node_metrics(self, db):
+    """
+    Get historical monthly metrics and last month's daily metrics
+    to periodically synchronize metrics across all nodes
+    """
     daily_route_metrics = {}
     monthly_route_metrics = {}
     daily_app_metrics = {}
@@ -392,55 +263,15 @@ def synchronize_all_node_metrics(self, db):
 
 # ####### CELERY TASKS ####### #
 
-
-@celery.task(name="update_metrics", bind=True)
-@save_duration_metric(metric_group="celery_task")
-def update_metrics(self):
-    # Cache custom task class properties
-    # Details regarding custom task context can be found in wiki
-    # Custom Task definition can be found in src/app.py
-    db = update_metrics.db
-    redis = update_metrics.redis
-
-    # Define lock acquired boolean
-    have_lock = False
-
-    # Define redis lock object
-    update_lock = redis.lock("update_metrics_lock", blocking_timeout=25)
-    try:
-        # Attempt to acquire lock - do not block if unable to acquire
-        have_lock = update_lock.acquire(blocking=False)
-        if have_lock:
-            logger.info(
-                f"index_metrics.py | update_metrics | {self.request.id} | Acquired update_metrics_lock"
-            )
-            metric = PrometheusMetric(
-                PrometheusMetricNames.INDEX_METRICS_DURATION_SECONDS
-            )
-            sweep_metrics(db, redis)
-            refresh_metrics_matviews(db)
-            metric.save_time({"task_name": "update_metrics"})
-            logger.info(
-                f"index_metrics.py | update_metrics | {self.request.id} | Processing complete within session"
-            )
-        else:
-            logger.error(
-                f"index_metrics.py | update_metrics | {self.request.id} | Failed to acquire update_metrics_lock"
-            )
-    except Exception as e:
-        logger.error("Fatal error in main loop of update_metrics: %s", e, exc_info=True)
-        raise e
-    finally:
-        if have_lock:
-            update_lock.release()
-
-
 @celery.task(name="aggregate_metrics", bind=True)
 @save_duration_metric(metric_group="celery_task")
 def aggregate_metrics(self):
-    # Cache custom task class properties
-    # Details regarding custom task context can be found in wiki
-    # Custom Task definition can be found in src/app.py
+    """
+    Aggregates metrics from other nodes.
+
+    Get recent route and app metrics from all other discovery nodes (ip, app name, etc)
+    and consolidates (deduplicating ips) with this node's metrics.
+    """
     db = aggregate_metrics.db
     redis = aggregate_metrics.redis
 
@@ -481,9 +312,13 @@ def aggregate_metrics(self):
 @celery.task(name="synchronize_metrics", bind=True)
 @save_duration_metric(metric_group="celery_task")
 def synchronize_metrics(self):
-    # Cache custom task class properties
-    # Details regarding custom task context can be found in wiki
-    # Custom Task definition can be found in src/app.py
+    """
+    Ensures that metrics reporting is consistent between all nodes.
+
+    Gets historical metrics from all other discovery nodes unique/total counts
+    for requests and app names and updates this node with the max(self, other) to
+    ensure that each discovery node reports metrics consistently.
+    """
     db = synchronize_metrics.db
     redis = synchronize_metrics.redis
